move files in models package
